Restructure Scripts to have UI section
Also restructure UIControllers to be inheritance-based

Also rehome some poor lonely root-folder scripts
