+++ Assets/Scripts/Consequences/BeginDroneLaunchConsequence.cs
