+++ Assets/Scripts/TriggeredEffects/ActivateRigidbodyOnTrigger.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
-using Player.PlayerStatusEffect;
-using PlayerStatusEffect;
+using UI;
 using UnityEngine;
 
 public class ActivateRigidbodyOnTrigger : MonoBehaviour
