+++ Assets/Scripts/UI/IUIController.cs
@@ -1,0 +1,7 @@
+namespace UI
+{
+    public interface IUIController
+    {
+        public void ShowNextFrame();
+    }
+}
