+++ Assets/Scripts/UI/IUIController.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f2f6adb663234bd5a40715e244ed27dc
+timeCreated: 1683403827
