+++ Assets/Scripts/UI/ScoreScreenUIController.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace UI
+{
+    public class ScoreScreenUIController : MonoBehaviour
+    {
+        
+    }
+}
