+++ Assets/Scripts/UI/ScoreScreenUIController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1785c7449f0744145a304de752dd0263
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
