+++ Assets/Scripts/UI/StatusEffectUIController.cs
@@ -1,0 +1,10 @@
+using TMPro;
+using UnityEngine;
+
+namespace UI
+{
+    public class StatusEffectUIController : TextUIController
+    {
+        public string statusName;
+    }
+}
