+++ Assets/Scripts/UI/TextUIController.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7e63c615d5804addba27e9941ed88582
+timeCreated: 1683403789
